+++ ozhera-monitor/ozhera-monitor-service/src/main/java/org/apache/ozhera/monitor/service/HeraBaseInfoService.java
@@ -34,7 +34,7 @@ public interface HeraBaseInfoService {
     
     Result addRole(HeraAppRoleModel model);
     
-    Result delRole(Integer id);
+    Result delRole(Long id);
     
     Result queryRole(HeraAppRoleModel model, Integer pageNo, Integer pageCount);
